requirements.txt
pandas==1.1.3
pytesseract==0.3.6
image==1.5.33
python_docx==0.8.10
docx==0.2.4
Pillow==8.0.1
PyPDF2==1.26.0